w3c-reports/12092014/efl/runned-tests/dom/nodes/Document-characterSet-normalization-expected.txt


PASS Name "utf-8" has label "unicode-1-1-utf-8" 
PASS Name "utf-8" has label "utf-8" 
PASS Name "utf-8" has label "utf8" 
PASS Name "ibm866" has label "866" 
PASS Name "ibm866" has label "cp866" 
PASS Name "ibm866" has label "csibm866" 
PASS Name "ibm866" has label "ibm866" 
PASS Name "iso-8859-2" has label "csisolatin2" 
PASS Name "iso-8859-2" has label "iso-8859-2" 
PASS Name "iso-8859-2" has label "iso-ir-101" 
PASS Name "iso-8859-2" has label "iso8859-2" 
FAIL Name "iso-8859-2" has label "iso88592" assert_equals: expected "iso-8859-2" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-2" has label "iso_8859-2" assert_equals: expected "iso-8859-2" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-2" has label "iso_8859-2:1987" 
PASS Name "iso-8859-2" has label "l2" 
PASS Name "iso-8859-2" has label "latin2" 
PASS Name "iso-8859-3" has label "csisolatin3" 
PASS Name "iso-8859-3" has label "iso-8859-3" 
PASS Name "iso-8859-3" has label "iso-ir-109" 
PASS Name "iso-8859-3" has label "iso8859-3" 
FAIL Name "iso-8859-3" has label "iso88593" assert_equals: expected "iso-8859-3" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-3" has label "iso_8859-3" assert_equals: expected "iso-8859-3" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-3" has label "iso_8859-3:1988" 
PASS Name "iso-8859-3" has label "l3" 
PASS Name "iso-8859-3" has label "latin3" 
PASS Name "iso-8859-4" has label "csisolatin4" 
PASS Name "iso-8859-4" has label "iso-8859-4" 
PASS Name "iso-8859-4" has label "iso-ir-110" 
PASS Name "iso-8859-4" has label "iso8859-4" 
FAIL Name "iso-8859-4" has label "iso88594" assert_equals: expected "iso-8859-4" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-4" has label "iso_8859-4" assert_equals: expected "iso-8859-4" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-4" has label "iso_8859-4:1988" 
PASS Name "iso-8859-4" has label "l4" 
PASS Name "iso-8859-4" has label "latin4" 
PASS Name "iso-8859-5" has label "csisolatincyrillic" 
PASS Name "iso-8859-5" has label "cyrillic" 
PASS Name "iso-8859-5" has label "iso-8859-5" 
PASS Name "iso-8859-5" has label "iso-ir-144" 
PASS Name "iso-8859-5" has label "iso8859-5" 
FAIL Name "iso-8859-5" has label "iso88595" assert_equals: expected "iso-8859-5" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-5" has label "iso_8859-5" assert_equals: expected "iso-8859-5" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-5" has label "iso_8859-5:1988" 
PASS Name "iso-8859-6" has label "arabic" 
PASS Name "iso-8859-6" has label "asmo-708" 
FAIL Name "iso-8859-6" has label "csiso88596e" assert_equals: expected "iso-8859-6" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-6" has label "csiso88596i" assert_equals: expected "iso-8859-6" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-6" has label "csisolatinarabic" 
PASS Name "iso-8859-6" has label "ecma-114" 
PASS Name "iso-8859-6" has label "iso-8859-6" 
PASS Name "iso-8859-6" has label "iso-8859-6-e" 
PASS Name "iso-8859-6" has label "iso-8859-6-i" 
PASS Name "iso-8859-6" has label "iso-ir-127" 
PASS Name "iso-8859-6" has label "iso8859-6" 
FAIL Name "iso-8859-6" has label "iso88596" assert_equals: expected "iso-8859-6" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-6" has label "iso_8859-6" assert_equals: expected "iso-8859-6" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-6" has label "iso_8859-6:1987" 
PASS Name "iso-8859-7" has label "csisolatingreek" 
PASS Name "iso-8859-7" has label "ecma-118" 
PASS Name "iso-8859-7" has label "elot_928" 
PASS Name "iso-8859-7" has label "greek" 
PASS Name "iso-8859-7" has label "greek8" 
PASS Name "iso-8859-7" has label "iso-8859-7" 
PASS Name "iso-8859-7" has label "iso-ir-126" 
PASS Name "iso-8859-7" has label "iso8859-7" 
FAIL Name "iso-8859-7" has label "iso88597" assert_equals: expected "iso-8859-7" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-7" has label "iso_8859-7" assert_equals: expected "iso-8859-7" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-7" has label "iso_8859-7:1987" 
PASS Name "iso-8859-7" has label "sun_eu_greek" 
FAIL Name "iso-8859-8" has label "csiso88598e" assert_equals: expected "iso-8859-8" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-8" has label "csisolatinhebrew" 
PASS Name "iso-8859-8" has label "hebrew" 
PASS Name "iso-8859-8" has label "iso-8859-8" 
PASS Name "iso-8859-8" has label "iso-8859-8-e" 
PASS Name "iso-8859-8" has label "iso-ir-138" 
PASS Name "iso-8859-8" has label "iso8859-8" 
FAIL Name "iso-8859-8" has label "iso88598" assert_equals: expected "iso-8859-8" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-8" has label "iso_8859-8" assert_equals: expected "iso-8859-8" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-8" has label "iso_8859-8:1988" 
PASS Name "iso-8859-8" has label "visual" 
PASS Name "iso-8859-8-i" has label "csiso88598i" 
PASS Name "iso-8859-8-i" has label "iso-8859-8-i" 
PASS Name "iso-8859-8-i" has label "logical" 
PASS Name "iso-8859-10" has label "csisolatin6" 
PASS Name "iso-8859-10" has label "iso-8859-10" 
PASS Name "iso-8859-10" has label "iso-ir-157" 
PASS Name "iso-8859-10" has label "iso8859-10" 
FAIL Name "iso-8859-10" has label "iso885910" assert_equals: expected "iso-8859-10" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-10" has label "l6" 
PASS Name "iso-8859-10" has label "latin6" 
PASS Name "iso-8859-13" has label "iso-8859-13" 
PASS Name "iso-8859-13" has label "iso8859-13" 
FAIL Name "iso-8859-13" has label "iso885913" assert_equals: expected "iso-8859-13" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-14" has label "iso-8859-14" 
PASS Name "iso-8859-14" has label "iso8859-14" 
FAIL Name "iso-8859-14" has label "iso885914" assert_equals: expected "iso-8859-14" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-15" has label "csisolatin9" 
PASS Name "iso-8859-15" has label "iso-8859-15" 
PASS Name "iso-8859-15" has label "iso8859-15" 
FAIL Name "iso-8859-15" has label "iso885915" assert_equals: expected "iso-8859-15" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "iso-8859-15" has label "iso_8859-15" assert_equals: expected "iso-8859-15" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "iso-8859-15" has label "l9" 
FAIL Name "iso-8859-16" has label "iso-8859-16" assert_equals: expected "iso-8859-16" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "koi8-r" has label "cskoi8r" 
PASS Name "koi8-r" has label "koi" 
PASS Name "koi8-r" has label "koi8" 
PASS Name "koi8-r" has label "koi8-r" 
FAIL Name "koi8-r" has label "koi8_r" assert_equals: expected "koi8-r" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "koi8-u" has label "koi8-u" 
PASS Name "macintosh" has label "csmacintosh" 
PASS Name "macintosh" has label "mac" 
PASS Name "macintosh" has label "macintosh" 
PASS Name "macintosh" has label "x-mac-roman" 
PASS Name "windows-874" has label "dos-874" 
PASS Name "windows-874" has label "iso-8859-11" 
PASS Name "windows-874" has label "iso8859-11" 
FAIL Name "windows-874" has label "iso885911" assert_equals: expected "windows-874" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-874" has label "tis-620" 
PASS Name "windows-874" has label "windows-874" 
PASS Name "windows-1250" has label "cp1250" 
PASS Name "windows-1250" has label "windows-1250" 
PASS Name "windows-1250" has label "x-cp1250" 
PASS Name "windows-1251" has label "cp1251" 
PASS Name "windows-1251" has label "windows-1251" 
PASS Name "windows-1251" has label "x-cp1251" 
FAIL Name "windows-1252" has label "ansi_x3.4-1968" assert_equals: expected "windows-1252" but got "us-ascii"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "ascii" assert_equals: expected "windows-1252" but got "us-ascii"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1252" has label "cp1252" 
FAIL Name "windows-1252" has label "cp819" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "csisolatin1" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "ibm819" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "iso-8859-1" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "iso-ir-100" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "iso8859-1" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "iso88591" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "iso_8859-1" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "iso_8859-1:1987" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "l1" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "latin1" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1252" has label "us-ascii" assert_equals: expected "windows-1252" but got "us-ascii"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1252" has label "windows-1252" 
FAIL Name "windows-1252" has label "x-cp1252" assert_equals: expected "windows-1252" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1253" has label "cp1253" 
PASS Name "windows-1253" has label "windows-1253" 
FAIL Name "windows-1253" has label "x-cp1253" assert_equals: expected "windows-1253" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1254" has label "cp1254" 
PASS Name "windows-1254" has label "csisolatin5" 
PASS Name "windows-1254" has label "iso-8859-9" 
PASS Name "windows-1254" has label "iso-ir-148" 
PASS Name "windows-1254" has label "iso8859-9" 
FAIL Name "windows-1254" has label "iso88599" assert_equals: expected "windows-1254" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
FAIL Name "windows-1254" has label "iso_8859-9" assert_equals: expected "windows-1254" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1254" has label "iso_8859-9:1989" 
PASS Name "windows-1254" has label "l5" 
PASS Name "windows-1254" has label "latin5" 
PASS Name "windows-1254" has label "windows-1254" 
FAIL Name "windows-1254" has label "x-cp1254" assert_equals: expected "windows-1254" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1255" has label "cp1255" 
PASS Name "windows-1255" has label "windows-1255" 
FAIL Name "windows-1255" has label "x-cp1255" assert_equals: expected "windows-1255" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1256" has label "cp1256" 
PASS Name "windows-1256" has label "windows-1256" 
FAIL Name "windows-1256" has label "x-cp1256" assert_equals: expected "windows-1256" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1257" has label "cp1257" 
PASS Name "windows-1257" has label "windows-1257" 
FAIL Name "windows-1257" has label "x-cp1257" assert_equals: expected "windows-1257" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "windows-1258" has label "cp1258" 
PASS Name "windows-1258" has label "windows-1258" 
FAIL Name "windows-1258" has label "x-cp1258" assert_equals: expected "windows-1258" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "x-mac-cyrillic" has label "x-mac-cyrillic" 
PASS Name "x-mac-cyrillic" has label "x-mac-ukrainian" 
PASS Name "gbk" has label "chinese" 
PASS Name "gbk" has label "csgb2312" 
PASS Name "gbk" has label "csiso58gb231280" 
PASS Name "gbk" has label "gb2312" 
FAIL Name "gbk" has label "gb_2312" assert_equals: expected "gbk" but got "iso-8859-1"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "gbk" has label "gb_2312-80" 
PASS Name "gbk" has label "gbk" 
PASS Name "gbk" has label "iso-ir-58" 
PASS Name "gbk" has label "x-gbk" 
PASS Name "gb18030" has label "gb18030" 
PASS Name "hz-gb-2312" has label "hz-gb-2312" 
PASS Name "big5" has label "big5" 
FAIL Name "big5" has label "big5-hkscs" assert_equals: expected "big5" but got "big5-hkscs"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
PASS Name "big5" has label "cn-big5" 
PASS Name "big5" has label "csbig5" 
PASS Name "big5" has label "x-x-big5" 
PASS Name "euc-jp" has label "cseucpkdfmtjapanese" 
PASS Name "euc-jp" has label "euc-jp" 
PASS Name "euc-jp" has label "x-euc-jp" 
PASS Name "iso-2022-jp" has label "csiso2022jp" 
PASS Name "iso-2022-jp" has label "iso-2022-jp" 
PASS Name "shift_jis" has label "csshiftjis" 
PASS Name "shift_jis" has label "ms_kanji" 
PASS Name "shift_jis" has label "shift-jis" 
PASS Name "shift_jis" has label "shift_jis" 
PASS Name "shift_jis" has label "sjis" 
PASS Name "shift_jis" has label "windows-31j" 
PASS Name "shift_jis" has label "x-sjis" 
PASS Name "euc-kr" has label "cseuckr" 
PASS Name "euc-kr" has label "csksc56011987" 
PASS Name "euc-kr" has label "euc-kr" 
PASS Name "euc-kr" has label "iso-ir-149" 
PASS Name "euc-kr" has label "korean" 
PASS Name "euc-kr" has label "ks_c_5601-1987" 
PASS Name "euc-kr" has label "ks_c_5601-1989" 
PASS Name "euc-kr" has label "ksc5601" 
PASS Name "euc-kr" has label "ksc_5601" 
PASS Name "euc-kr" has label "windows-949" 
PASS Name "iso-2022-kr" has label "csiso2022kr" 
PASS Name "iso-2022-kr" has label "iso-2022-kr" 
FAIL Name "x-user-defined" has label "x-user-defined" assert_equals: expected "x-user-defined" but got "windows-1252"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:358:22
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/dom/nodes/Document-characterSet-normalization.html:357:13)
